Multiple changes to improve library

+++ src/AT24C256.h
@@ -8,12 +8,12 @@
 class AT24C256
 {
 public: 
-		AT24C256(int address);
+		AT24C256(uint8_t address);
 		void write(uint16_t writeAddress, uint8_t* data, uint8_t len);
 		void read(uint16_t readAddress, uint8_t* data, uint8_t len);
 
 private:
-		int _address;
+		unsigned short int _address;
 
 };
